server/requirements.txt
asgiref==3.7.2
astroid==3.0.2
asttokens==2.4.1
boto3==1.34.55
botocore==1.34.55
cfgv==3.4.0
decorator==5.1.1
dill==0.3.7
distlib==0.3.8
django-cors-headers==4.3.1
django-extensions==3.2.3
django-storages==1.14.2
Django==5.0
djangorestframework-simplejwt==5.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
executing==2.0.1
filelock==3.13.1
gunicorn==21.2.0
identify==2.5.34
inflection==0.5.1
ipython==8.18.1
isort==5.13.2
jedi==0.19.1
jmespath==1.0.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mysql-connector-python==8.2.0
mysqlclient==2.2.4
nodeenv==1.8.0
numpy
packaging==23.2
parso==0.8.3
pep8==1.7.1
pexpect==4.9.0
platformdirs==4.1.0
pre-commit==3.6.1
prompt-toolkit==3.0.43
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pylint==3.0.3
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomlkit==0.12.3
traitlets==5.14.0
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.25.0
wcwidth==0.2.12
whitenoise==6.6.0